requirements.txt
google-api-python-client==2.149.0
google-auth-oauthlib==1.2.1
google-auth==2.35.0
python-dotenv==1.0.1
requests==2.32.3
pytesseract==0.3.13
Pillow==10.4.0
pdf2image==1.17.0
